# Update Hubbard Model & LatPhys compatability