--- a/Project.toml
+++ b/Project.toml
@@ -29,7 +29,7 @@ Reexport = "0.2"
 Requires = "0.5, 1.0"
 StableDQMC = "0.1"
 TimerOutputs = "0.5.3"
-julia = "1"
+julia = "1.2"
 
 [extras]
 InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"