--- a/src/helpers.jl
+++ b/src/helpers.jl
@@ -96,7 +96,18 @@ function SparseArrays.mul!(C::StridedMatrix, X::StridedMatrix, A::SparseMatrixCS
     C
 end
 
-
+# LoopVectorization
+# Let's assume equal size
+# See LoopVectorization.jl
+function vmul!(C, A, B)
+    @avx for m in axes(A,1), n in axes(B,2)
+        Cmn = zero(eltype(C))
+        for k in axes(A,2)
+            Cmn += A[m,k] * B[k,n]
+        end
+        C[m,n] = Cmn
+    end
+end
 
 # Taken from Base
 if !isdefined(Base, :splitpath)