--- a/src/models/HubbardAttractive/HubbardModelAttractive.jl
+++ b/src/models/HubbardAttractive/HubbardModelAttractive.jl
@@ -30,6 +30,9 @@ with linear system size `L`. Additional allowed `kwargs` are:
     l::LT = choose_lattice(HubbardModelAttractive, dims, L)
     neighs::Matrix{Int} = neighbors_lookup_table(l)
     flv::Int = 1
+    # to avoid allocations (TODO always real?)
+    IG::Vector{Float64} = Vector{Float64}(undef, length(l))
+    G::Vector{Float64} = Vector{Float64}(undef, length(l))
 end
 
 
@@ -140,10 +143,26 @@ end
     greens = mc.s.greens
     γ = delta
 
-    u = -greens[:, i]
-    u[i] += 1.
-    # TODO: OPT: speed check, maybe @views/@inbounds
-    greens .-= kron(u * 1. /(1 + γ * u[i]), transpose(γ * greens[i, :]))
+    # Unoptimized Version
+    # u = -greens[:, i]
+    # u[i] += 1.
+    # # OPT: speed check, maybe @views/@inbounds
+    # greens .-= kron(u * 1./(1 + gamma * u[i]), transpose(gamma * greens[i, :]))
+    # conf[i, slice] .*= -1.
+
+    # Optimized
+    # copy! and `.=` allocate, this doesn't. Synced loop is marginally faster
+    @inbounds for j in eachindex(m.IG)
+        m.IG[j] = -greens[j, i]
+        m.G[j] = greens[i, j]
+    end
+    @inbounds m.IG[i] += 1.0
+    # This is way faster for small systems and still ~33% faster at L = 15
+    # Also no allocations here
+    x = γ / (1.0 + γ * m.IG[i])
+    @inbounds for k in eachindex(m.IG), l in eachindex(m.G)
+        greens[k, l] -= m.IG[k] * x * m.G[l]
+    end
     conf[i, slice] *= -1
     nothing
 end