Shouldn't `interaction_matrix_exp!` include some term like `exp(-power * lambda * sum(conf[slice, ;]))`?