@ffreyer 2020-07-29
> Shouldn't `interaction_matrix_exp!` include some term like `exp(-power * lambda * sum(conf[slice, ;]))`?

No, that's given through `ΔE_boson` which `propose_local` returns.