@ffreyer 2020-07-30
julia 1.0 uses a really old version of LoopVectorization - that's why that fails. Julia 1.2 works, so does 1.4...
` Installed LoopVectorization ──── v0.1.3`